# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoScribe AI+ API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface SignupBody {
  fullName: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  fullName: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface Chat {
  id: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  guestSessionId?: string | null;
  title: string;
  mode: string;
  template: string;
  isSaved: boolean;
  /** @nullable */
  shareToken: string | null;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface Message {
  id: number;
  chatId: number;
  role: string;
  content: string;
  mode: string;
  /** @nullable */
  keywords: string | null;
  createdAt: string;
}

export interface ChatWithMessages {
  id: number;
  title: string;
  mode: string;
  template: string;
  isSaved: boolean;
  /** @nullable */
  shareToken: string | null;
  /** @nullable */
  prompt?: string | null;
  /** @nullable */
  reportOutput?: string | null;
  /** @nullable */
  codeOutput?: string | null;
  /** @nullable */
  docsOutput?: string | null;
  /** @nullable */
  insightsOutput?: string | null;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface CreateChatBody {
  title: string;
  mode: string;
  /** @nullable */
  guestSessionId?: string | null;
}

export interface UpdateChatBody {
  title: string;
}

export interface GenerateBody {
  prompt: string;
  mode: string;
  template?: string;
  /** @nullable */
  chatId?: number | null;
  /** @nullable */
  fileContent?: string | null;
  /** @nullable */
  guestSessionId?: string | null;
}

export interface GenerateResult {
  report: string;
  code: string;
  docs: string;
  insights: string;
  keywords: string[];
  suggestions: string[];
  /** @nullable */
  messageId: number | null;
  /** @nullable */
  chatId: number | null;
  title: string;
}

export interface UploadResult {
  content: string;
  filename: string;
  fileId: number;
}

export interface FileItem {
  id: number;
  filename: string;
  fileType: string;
  createdAt: string;
}

export interface SavedOutput {
  id: number;
  title: string;
  content: string;
  mode: string;
  /** @nullable */
  chatId: number | null;
  createdAt: string;
}

export interface CreateSavedOutputBody {
  title: string;
  content: string;
  mode: string;
  /** @nullable */
  chatId?: number | null;
  /** @nullable */
  guestSessionId?: string | null;
}

export interface ShareLink {
  token: string;
  url: string;
}

export interface ShareLinkItem {
  id: number;
  token: string;
  title: string;
  mode: string;
  createdAt: string;
}

export interface CreateShareBody {
  content: string;
  title: string;
  mode: string;
  /** @nullable */
  chatId?: number | null;
  /** @nullable */
  guestSessionId?: string | null;
}

export interface SharedOutput {
  title: string;
  content: string;
  mode: string;
  createdAt: string;
}

export interface ModeCount {
  mode: string;
  count: number;
}

export interface StatsResult {
  totalChats: number;
  totalMessages: number;
  totalSaved: number;
  modeBreakdown: ModeCount[];
}

export type ListChatsParams = {
  guestSessionId?: string;
};

export type UploadFileBody = {
  file: Blob;
  /** @nullable */
  chatId?: number | null;
};

export type ListFilesParams = {
  guestSessionId?: string;
};

export type ListSavedOutputsParams = {
  guestSessionId?: string;
};

export type ListSharedLinksParams = {
  guestSessionId?: string;
};

export type GetStatsParams = {
  guestSessionId?: string;
};