# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: AutoScribe AI+ API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: chats
  - name: messages
  - name: saved
  - name: files
  - name: share
  - name: ai
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user from JWT
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized

  /chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: List chats for current user or guest session
      parameters:
        - in: query
          name: guestSessionId
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create a new chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatBody"
      responses:
        "201":
          description: Created chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

  /chats/{id}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get a single chat with messages
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatWithMessages"
        "404":
          description: Not found
    patch:
      operationId: updateChat
      tags: [chats]
      summary: Update chat title
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChatBody"
      responses:
        "200":
          description: Updated chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
    delete:
      operationId: deleteChat
      tags: [chats]
      summary: Delete a chat
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /chats/{id}/save:
    patch:
      operationId: toggleSaveChat
      tags: [chats]
      summary: Toggle save/star on a chat
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

  /chats/{chatId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for a chat
      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  /generate:
    post:
      operationId: generate
      tags: [ai]
      summary: Generate AI response with structured output
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateBody"
      responses:
        "200":
          description: Generated structured result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResult"

  /upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file (TXT or CSV)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                chatId:
                  type: ["integer", "null"]
              required:
                - file
      responses:
        "200":
          description: File content extracted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"

  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List uploaded files for current user
      parameters:
        - in: query
          name: guestSessionId
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileItem"

  /saved:
    get:
      operationId: listSavedOutputs
      tags: [saved]
      summary: List saved outputs
      parameters:
        - in: query
          name: guestSessionId
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of saved outputs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedOutput"
    post:
      operationId: createSavedOutput
      tags: [saved]
      summary: Save an output
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavedOutputBody"
      responses:
        "201":
          description: Saved output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedOutput"

  /saved/{id}:
    delete:
      operationId: deleteSavedOutput
      tags: [saved]
      summary: Delete a saved output
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /share:
    post:
      operationId: createShareLink
      tags: [share]
      summary: Create a share link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShareBody"
      responses:
        "201":
          description: Share link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLink"

  /share/list:
    get:
      operationId: listSharedLinks
      tags: [share]
      summary: List share links for current user
      parameters:
        - in: query
          name: guestSessionId
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of share links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShareLinkItem"

  /share/{token}:
    get:
      operationId: getSharedOutput
      tags: [share]
      summary: Get a shared output by token (public)
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shared output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedOutput"
        "404":
          description: Not found

  /stats:
    get:
      operationId: getStats
      tags: [chats]
      summary: Get usage statistics
      parameters:
        - in: query
          name: guestSessionId
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    SignupBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
      required: [fullName, email, password]

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        createdAt:
          type: string
      required: [id, fullName, email, createdAt]

    Chat:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        guestSessionId:
          type: ["string", "null"]
        title:
          type: string
        mode:
          type: string
        template:
          type: string
        isSaved:
          type: boolean
        shareToken:
          type: ["string", "null"]
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, mode, template, isSaved, shareToken, messageCount, createdAt, updatedAt]

    ChatWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        mode:
          type: string
        template:
          type: string
        isSaved:
          type: boolean
        shareToken:
          type: ["string", "null"]
        prompt:
          type: ["string", "null"]
        reportOutput:
          type: ["string", "null"]
        codeOutput:
          type: ["string", "null"]
        docsOutput:
          type: ["string", "null"]
        insightsOutput:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required: [id, title, mode, template, isSaved, shareToken, createdAt, updatedAt, messages]

    CreateChatBody:
      type: object
      properties:
        title:
          type: string
        mode:
          type: string
        guestSessionId:
          type: ["string", "null"]
      required: [title, mode]

    UpdateChatBody:
      type: object
      properties:
        title:
          type: string
      required: [title]

    Message:
      type: object
      properties:
        id:
          type: integer
        chatId:
          type: integer
        role:
          type: string
        content:
          type: string
        mode:
          type: string
        keywords:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, chatId, role, content, mode, keywords, createdAt]

    GenerateBody:
      type: object
      properties:
        prompt:
          type: string
        mode:
          type: string
        template:
          type: string
        chatId:
          type: ["integer", "null"]
        fileContent:
          type: ["string", "null"]
        guestSessionId:
          type: ["string", "null"]
      required: [prompt, mode]

    GenerateResult:
      type: object
      properties:
        report:
          type: string
        code:
          type: string
        docs:
          type: string
        insights:
          type: string
        keywords:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        messageId:
          type: ["integer", "null"]
        chatId:
          type: ["integer", "null"]
        title:
          type: string
      required: [report, code, docs, insights, keywords, suggestions, messageId, chatId, title]

    UploadResult:
      type: object
      properties:
        content:
          type: string
        filename:
          type: string
        fileId:
          type: integer
      required: [content, filename, fileId]

    FileItem:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        fileType:
          type: string
        createdAt:
          type: string
      required: [id, filename, fileType, createdAt]

    SavedOutput:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        mode:
          type: string
        chatId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, title, content, mode, chatId, createdAt]

    CreateSavedOutputBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        mode:
          type: string
        chatId:
          type: ["integer", "null"]
        guestSessionId:
          type: ["string", "null"]
      required: [title, content, mode]

    ShareLink:
      type: object
      properties:
        token:
          type: string
        url:
          type: string
      required: [token, url]

    ShareLinkItem:
      type: object
      properties:
        id:
          type: integer
        token:
          type: string
        title:
          type: string
        mode:
          type: string
        createdAt:
          type: string
      required: [id, token, title, mode, createdAt]

    CreateShareBody:
      type: object
      properties:
        content:
          type: string
        title:
          type: string
        mode:
          type: string
        chatId:
          type: ["integer", "null"]
        guestSessionId:
          type: ["string", "null"]
      required: [content, title, mode]

    SharedOutput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        mode:
          type: string
        createdAt:
          type: string
      required: [title, content, mode, createdAt]

    StatsResult:
      type: object
      properties:
        totalChats:
          type: integer
        totalMessages:
          type: integer
        totalSaved:
          type: integer
        modeBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ModeCount"
      required: [totalChats, totalMessages, totalSaved, modeBreakdown]

    ModeCount:
      type: object
      properties:
        mode:
          type: string
        count:
          type: integer
      required: [mode, count]