# lib/api-zod/src/generated/types/createShareBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoScribe AI+ API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreateShareBody {
  content: string;
  title: string;
  mode: string;
  /** @nullable */
  chatId?: number | null;
  /** @nullable */
  guestSessionId?: string | null;
}