# lib/api-zod/src/generated/types/fileItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoScribe AI+ API specification
 * OpenAPI spec version: 0.2.0
 */

export interface FileItem {
  id: number;
  filename: string;
  fileType: string;
  createdAt: string;
}