# lib/api-zod/src/generated/types/generateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoScribe AI+ API specification
 * OpenAPI spec version: 0.2.0
 */

export interface GenerateBody {
  prompt: string;
  mode: string;
  template?: string;
  /** @nullable */
  chatId?: number | null;
  /** @nullable */
  fileContent?: string | null;
  /** @nullable */
  guestSessionId?: string | null;
}