# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoScribe AI+ API specification
 * OpenAPI spec version: 0.2.0
 */

export interface SignupBody {
  fullName: string;
  email: string;
  password: string;
}